Test for get rewards est with pool incentives (#329)
* Test for get rewards EST with pool incentives

* Undo state breaking change
